Move code and add doc.
